move to other kzip directory
