+++ kythe/go/platform/tools/kzip/infocmd/BUILD
@@ -8,22 +8,11 @@ go_library(
         "infocmd.go",
     ],
     deps = [
-        ":info",
+        "//kythe/go/platform/kzip/info",
         "//kythe/go/platform/vfs",
         "//kythe/go/util/cmdutil",
         "@com_github_golang_protobuf//jsonpb:go_default_library_gen",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_google_subcommands//:go_default_library",
-    ],
-)
-
-go_library(
-    name = "info",
-    srcs = [
-        "info.go",
-    ],
-    deps = [
-        "//kythe/go/platform/kzip",
-        "//kythe/proto:analysis_go_proto",
     ],
 )
